+++ packages/design-tokens/src/__tests__/sass-tokens.spec.ts
@@ -2,34 +2,42 @@ import fs from "fs"
 import path from "path"
 import PostCSS, { Declaration } from "postcss"
 
-import { heartColorNamePattern } from "../patterns"
+import { heartColorNamePattern } from "../lib/patterns"
 
 const readSassFile = (fileName: string) =>
   fs.readFileSync(path.join(`${__dirname}../../../sass/`, fileName))
 
 describe("SASS tokens", () => {
+  const colors = PostCSS.parse(readSassFile("color.scss"))
+  const layout = PostCSS.parse(readSassFile("layout.scss"))
+  const shadow = PostCSS.parse(readSassFile("shadow.scss"))
+  const border = PostCSS.parse(readSassFile("border.scss"))
+  const spacing = PostCSS.parse(readSassFile("spacing.scss"))
+  const animation = PostCSS.parse(readSassFile("animation.scss"))
+  const typography = PostCSS.parse(readSassFile("typography.scss"))
+  const everything = PostCSS.parse(
+    `
+${colors.toString()}
+${layout.toString()}
+${shadow.toString()}
+${border.toString()}
+${spacing.toString()}
+${animation.toString()}
+${typography.toString()}
+    `
+  )
+  describe("everything", () => {
+    test("no tokens start with kz", () => {
+      const allTokensStartingWithKz = everything.nodes
+        .filter(
+          (n): n is Declaration => n.type === "decl" && n.prop.startsWith("$kz")
+        )
+        .map(n => n.prop)
+      expect(allTokensStartingWithKz).toHaveLength(0)
+    })
+  })
+
   describe("colors", () => {
-    const colors = PostCSS.parse(readSassFile("color.scss"))
-    const colorVars = PostCSS.parse(readSassFile("color-vars.scss"))
-    test("kz-var tokens don't contain heart color names", () => {
-      expect(
-        colorVars.nodes.filter(
-          n => n.type === "decl" && heartColorNamePattern.test(n.prop)
-        )
-      ).toHaveLength(0)
-    })
-
-    test("kz tokens don't contain heart color names", () => {
-      expect(
-        colors.nodes.filter(
-          n =>
-            n.type === "decl" &&
-            n.prop.startsWith("$kz") &&
-            heartColorNamePattern.test(n.prop)
-        )
-      ).toHaveLength(0)
-    })
-
     test("new un-prefixed tokens only contain heart color names", () => {
       const newColors = colors.nodes.filter(
         (n): n is Declaration =>
@@ -46,16 +54,6 @@ describe("SASS tokens", () => {
   })
 
   describe("layout", () => {
-    const layout = PostCSS.parse(readSassFile("layout.scss"))
-    const layoutVars = PostCSS.parse(readSassFile("layout-vars.scss"))
-    test("kz-var tokens are all CSS variables", () => {
-      expect(
-        layoutVars.nodes.filter(
-          n => n.type === "decl" && !n.value.startsWith("var(")
-        )
-      ).toHaveLength(0)
-    })
-
     test("new un-prefixed layout tokens are not CSS variables", () => {
       expect(
         layout.nodes.filter(