+++ draft-packages/empty-state/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.5.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-empty-state@2.4.1...@kaizen/draft-empty-state@2.5.0) (2021-08-13)
+
+
+### Features
+
+* Bump design-tokens in peer dependencies to v3 ([#1840](https://github.com/cultureamp/kaizen-design-system/issues/1840)) ([ca45bf4](https://github.com/cultureamp/kaizen-design-system/commit/ca45bf4707b5fbf907163653549e17682c46f636))
+
+
+
+
+
+## [2.4.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-empty-state@2.4.0...@kaizen/draft-empty-state@2.4.1) (2021-08-12)
+
+**Note:** Version bump only for package @kaizen/draft-empty-state
+
+
+
+
+
+# [2.4.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-empty-state@2.3.25...@kaizen/draft-empty-state@2.4.0) (2021-08-11)
+
+
+### Features
+
+* Migrate scss to use new design-tokens ([#1813](https://github.com/cultureamp/kaizen-design-system/issues/1813)) ([ec777a3](https://github.com/cultureamp/kaizen-design-system/commit/ec777a306cec1988894a9518b43f5247d500aa7d))
+
+
+
+
+
 ## [2.3.25](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-empty-state@2.3.24...@kaizen/draft-empty-state@2.3.25) (2021-08-10)
 
 **Note:** Version bump only for package @kaizen/draft-empty-state